@@ -1,60 +1,45 @@
 {
   "name": "guiatv",
   "version": "0.0.0",
+  "private": true,
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "start:ssr": "ng serve --port 4200",
     "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "dev:ssr": "ng run guiatv:serve-ssr",
-    "serve:ssr": "node dist/guiatv/server/server.js",
     "build:ssr": "ng build",
-    "prerender": "ng run guiatv:prerender"
+    "serve:ssr": "node dist/guiatv/server/server.mjs",
+    "dev:ssr": "ng serve --port 4200",
+    "prerender": "ng build --prerender"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/fire": "^17.1.0",
-    "@angular/forms": "^17.3.12",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/platform-server": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/ssr": "^17.3.12",
-    "bootstrap-icons": "^1.11.3",
+    "@angular/animations": "^20.0.0",
+    "@angular/cdk": "^20.0.1",
+    "@angular/common": "^20.0.0",
+    "@angular/compiler": "^20.0.0",
+    "@angular/core": "^20.0.0",
+    "@angular/forms": "^20.0.0",
+    "@angular/localize": "^20.0.0",
+    "@angular/material": "^20.0.1",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.0.0",
+    "@angular/platform-server": "^20.0.0",
+    "@angular/router": "^20.0.0",
+    "@angular/ssr": "^20.0.0",
+    "@types/leaflet": "^1.9.20",
+    "bootstrap-icons": "^1.13.1",
     "express": "^4.21.2",
-    "rxjs": "~7.8.1",
-    "swiper": "^11.2.8",
-    "tailwind-scrollbar-hide": "^1.1.7",
-    "tailwindcss-icons": "^1.0.2",
-    "tailwindcss-textshadow": "^2.1.3",
-    "tslib": "^2.6.0",
-    "zone.js": "~0.14.3"
+    "leaflet": "^1.9.4",
+    "rxjs": "~7.8.0",
+    "swiper": "^11.2.10",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler-cli": "^17.3.12",
-    "@types/express": "^4.17.22",
-    "@types/jasmine": "~5.1.4",
-    "@types/node": "^20.14.10",
-    "autoprefixer": "^10.4.19",
-    "browser-sync": "^3.0.4",
-    "jasmine-core": "~5.1.2",
-    "karma": "~6.4.3",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.1",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.4.39",
-    "tailwind-scrollbar": "^3.1.0",
-    "tailwindcss": "^3.4.4",
-    "typescript": "~5.4.5"
+    "@angular-devkit/build-angular": "^20.0.0",
+    "@angular/cli": "^20.0.0",
+    "@angular/compiler-cli": "^20.0.0",
+    "@types/express": "^4.17.17",
+    "@types/node": "^22.0.0",
+    "typescript": "~5.8.0"
   }
 }
