@@ -1,10 +1,13 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { ModalService } from 'src/app/services/modal.service';
+import { ModalService } from '../../services/modal.service';
 
 @Component({
   selector: 'app-modal',
   templateUrl: './modal.component.html',
   styleUrls: ['./modal.component.scss'],
+  standalone: true,
+  imports: [CommonModule],
 })
 export class ModalComponent {
   public program_modal: any = {};
