@@ -2,7 +2,6 @@ import { Component, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BlogDetailsComponent } from './blog-details.component';
 import { RouterModule, Routes } from '@angular/router';
-import { ComponentsModule } from 'src/app/components/components.module';
 
 const routes: Routes = [
   {
@@ -12,7 +11,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [BlogDetailsComponent],
-  imports: [CommonModule, RouterModule.forChild(routes), ComponentsModule],
+  declarations: [],
+  imports: [CommonModule, RouterModule.forChild(routes)],
 })
 export class BlogDetailsModule {}
