@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { ComponentsModule } from 'src/app/components/components.module';
 import { CanalCompletoComponent } from './canal-completo.component';
 import { FormsModule } from '@angular/forms';
 import { SliderComponent } from 'src/app/components/slider/slider.component';
@@ -14,10 +13,9 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [CanalCompletoComponent],
+  declarations: [],
   imports: [
     CommonModule,
-    ComponentsModule,
     RouterModule.forChild(routes),
     FormsModule,
     SliderComponent
