@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ListaCanalesComponent } from './lista-canales.component';
-import { ComponentsModule } from 'src/app/components/components.module';
 import { RouterModule, Routes } from '@angular/router';
 import { SliderComponent } from 'src/app/components/slider/slider.component';
 
@@ -13,7 +12,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [ListaCanalesComponent],
-  imports: [CommonModule, ComponentsModule, RouterModule.forChild(routes),SliderComponent],
+  declarations: [],
+  imports: [CommonModule, RouterModule.forChild(routes),SliderComponent],
 })
 export class ListaCanalesModule {}
