@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { ListaDestacadasComponent } from './lista-destacadas.component';
-import { ComponentsModule } from 'src/app/components/components.module';
 
 const routes: Routes = [
   {
@@ -12,7 +11,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [ListaDestacadasComponent],
-  imports: [CommonModule, ComponentsModule, RouterModule.forChild(routes)],
+  declarations: [],
+  imports: [CommonModule, RouterModule.forChild(routes)],
 })
 export class ListaDestacadasModule {}
