@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { PeliculasComponent } from './peliculas.component';
-import { ComponentsModule } from 'src/app/components/components.module';
 import { SliderComponent } from 'src/app/components/slider/slider.component';
 
 const routes: Routes = [
@@ -13,7 +12,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [PeliculasComponent],
-  imports: [CommonModule, RouterModule.forChild(routes), ComponentsModule,SliderComponent],
+  declarations: [],
+  imports: [CommonModule, RouterModule.forChild(routes),SliderComponent],
 })
 export class PeliculasModule {}
