@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Top10Component } from './top10.component';
 import { RouterModule, Routes } from '@angular/router';
-import { ComponentsModule } from 'src/app/components/components.module';
 import { SliderComponent } from 'src/app/components/slider/slider.component';
 
 const routes: Routes = [
@@ -13,11 +12,10 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [Top10Component],
+  declarations: [],
   imports: [
     CommonModule,
     RouterModule.forChild(routes),
-    ComponentsModule,
     SliderComponent
   ],
 })
