@@ -1,9 +1,9 @@
-//main.server.ts
+/// <reference types="@angular/localize" />
 
 import { bootstrapApplication } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
-import { appConfig } from './app/app.config.server';
+import { config } from './app/app.config.server';
 
-const bootstrap = () => bootstrapApplication(AppComponent, appConfig);
+const bootstrap = () => bootstrapApplication(AppComponent, config);
 
 export default bootstrap;
